Move the design-system primitives and platform shims out of Core/
Pure file moves. Core/ mixed design tokens, eleven view modifiers, window
plumbing, stores and pure algorithms in one 50-file namespace. Contents
unchanged; EdgeDissolve and ThinScrollbar are moved but not opened, per
AGENTS.md. KeyCapChip and IconCache are cut and pasted out of Theme.swift
and AppIndex.swift, byte-identical to their source hunks.

The two harness command lines that name a moved file are updated in the
same commit, in all four files that carry one: docs/development.md,
docs/snippets.md, checklists/testing.md and ci.yml.
