+++ .github/workflows/ci.yml
@@ -58,13 +58,13 @@ jobs:
           swiftc -swift-version 6 Tinycast/Core/CustomCommand.swift \
             Tinycast/Core/ShellCommandRunner.swift Tools/custom-command-test.swift \
             -o /tmp/custom-command-test && /tmp/custom-command-test
-          swiftc -swift-version 6 Tinycast/Core/NotificationToken.swift \
+          swiftc -swift-version 6 Tinycast/Platform/NotificationToken.swift \
             Tinycast/Core/Snippets/*.swift \
             Tools/snippets-test.swift -o /tmp/snippets-test && /tmp/snippets-test
           swiftc -swift-version 6 Tinycast/Core/HotKey/DoubleTapModifier.swift \
             Tinycast/Core/HotKey/DoubleTapDetector.swift Tools/hotkey-test.swift \
             -o /tmp/hotkey-test && /tmp/hotkey-test
-          swiftc -swift-version 6 Tinycast/Core/Theme.swift \
+          swiftc -swift-version 6 Tinycast/DesignSystem/Theme.swift \
             Tinycast/Core/CalloutPlacement.swift Tools/callout-test.swift \
             -o /tmp/callout-test && /tmp/callout-test
           swiftc -swift-version 6 Tinycast/Core/SystemAction.swift Tools/system-action-test.swift \
