+++ Tinycast/DesignSystem/KeyCapChip.swift
@@ -1,0 +1,55 @@
+import SwiftUI
+
+/// A single keycap chip: `.outline` for hotkey hints on rows, `.filled` for footer shortcuts.
+struct KeyCapChip: View {
+    enum Style {
+        case outline
+        case filled
+    }
+
+    /// Sanctioned cap sizes, so a bigger or smaller one is a named choice, not a stray frame.
+    enum Scale {
+        case compact
+        case standard
+        case hero
+
+        var side: CGFloat {
+            switch self {
+            case .compact: Theme.Size.compactKeyCap
+            case .standard: Theme.Size.keyCap
+            case .hero: Theme.Size.heroKeyCap
+            }
+        }
+
+        var font: Font {
+            switch self {
+            case .compact: Theme.Typography.compactKeyCap
+            case .standard: Theme.Typography.keyCap
+            case .hero: Theme.Typography.heroKeyCap
+            }
+        }
+    }
+
+    let text: String
+    var style: Style = .filled
+    var scale: Scale = .standard
+
+    /// "↵" is absent from SF Pro and falls back to Lucida Grande UI, which seats it 1.1pt higher in the line box than the SF caps — visibly top-heavy in a chip. Nudging via `offset` is render-only, so the chip keeps the same footprint as every other cap.
+    private static let returnGlyphDrop: CGFloat = 1.1
+
+    var body: some View {
+        let shape = RoundedRectangle(cornerRadius: Theme.Radius.keyCap, style: .continuous)
+        Text(text)
+            .font(scale.font)
+            .foregroundStyle(Theme.Colors.textSecondary)
+            .offset(y: text == "↵" ? Self.returnGlyphDrop : 0)
+            .padding(.horizontal, Theme.Spacing.xs)
+            .frame(minWidth: scale.side, minHeight: scale.side)
+            .background {
+                switch style {
+                case .filled: shape.fill(Theme.Colors.controlSurface)
+                case .outline: shape.strokeBorder(Theme.Colors.border, lineWidth: 1)
+                }
+            }
+    }
+}
