+++ Tinycast/Platform/AppDisplayName.swift
