+++ Tinycast/Platform/CursorScreen.swift
