+++ Tinycast/Platform/Images/ImageThumbnail.swift
