+++ Tinycast/Platform/LaunchAtLogin.swift
