+++ Tinycast/Platform/NotificationToken.swift
