+++ docs/development.md
@@ -96,13 +96,13 @@ swiftc Tinycast/Core/Emoji/EmojiCatalog.swift Tinycast/Core/Emoji/EmojiGridGeome
 swiftc -swift-version 6 Tinycast/Core/CustomCommand.swift \
     Tinycast/Core/ShellCommandRunner.swift Tools/custom-command-test.swift \
     -o /tmp/custom-command-test && /tmp/custom-command-test        # custom command store + runner
-swiftc -swift-version 6 Tinycast/Core/NotificationToken.swift \
+swiftc -swift-version 6 Tinycast/Platform/NotificationToken.swift \
     Tinycast/Core/HealthTicker.swift Tinycast/Core/Snippets/*.swift \
     Tools/snippets-test.swift -o /tmp/snippets-test && /tmp/snippets-test  # snippets
 swiftc -swift-version 6 Tinycast/Core/HotKey/DoubleTapModifier.swift \
     Tinycast/Core/HotKey/DoubleTapDetector.swift Tools/hotkey-test.swift \
     -o /tmp/hotkey-test && /tmp/hotkey-test                        # double-tap modifier recognizer
-swiftc -swift-version 6 Tinycast/Core/Theme.swift \
+swiftc -swift-version 6 Tinycast/DesignSystem/Theme.swift \
     Tinycast/Core/CalloutPlacement.swift Tools/callout-test.swift \
     -o /tmp/callout-test && /tmp/callout-test                      # shortcut-recorder callout placement
 swiftc -swift-version 6 Tinycast/Core/SystemAction.swift Tools/system-action-test.swift \
