+++ docs/refactor/ROADMAP.md
@@ -210,7 +210,7 @@ Update this table as phases land. `Blocked` requires a note in the phase's progr
 | 25  | Complete    | `refactor/25-remaining-coordinators`       | 2026-08-05 | **`AppCore` stops being the only orchestrator**: 1021 → 746 (−275), four coordinators, 2 → 6. All hard gates hold — `PaletteWindowController` is **absent from the diff** (so the frame, the session anchor and pop-to-root are byte-identical), `DialogController()` stays **1**, `AppCore.shared` greps empty in all four new files, and `runWindowCommand` stays put. Moves proven by direct body diff against `HEAD`, as in 24: every difference is a receiver rename or the one authorised `targetApp` helper, and the three pinned sequences (uninstall's ten steps, `perform`'s confirmation switch, `runCustomCommand`'s gate) are byte-identical. **`pickVolume` needed the third façade** `progress/24` predicted; `volumeHUD` and `auxWindows` did **not** — both moved ownership with their features instead. **AC2 not met — 746, not ~300, and the phase doc contradicts itself**: its own "net −370" budget lands at ~651. Of what remains, ~99 lines are the three palette types **phase 28 already extracts by name**, ~99 are forwarders **32** deletes, ~227 are composition root and things the phase says stay — and **~195 lines of palette-UI actions no phase owns**, which blocks 32's objective. Debug **and** Release clean, 0 → 0 warnings, all 17 harnesses green, `palette-selection-test` 111,684. **`regression.md` waived** — AC5–AC10 structural only, including the compact top-edge and held-hotkey checks the kickoff names. **Unblocks 26 and 32.** M4's DoD is **not** yet true; **25b** is proposed to close it. |
 | 25b | Complete    | `refactor/25b-palette-action-coordinators`  | 2026-08-06 | **M4's DoD is now literally true** — `AppCore` orchestrates no feature: 746 → 618 (−128), four coordinators, 6 → 10. Every hard gate holds — **`launch` is still one method** (`grep -c "func launch"` = 1) dispatching on `app.kind`, `PaletteWindowController` **and** `Paster` are both **absent from the diff**, `DialogController()` stays **1** with **no new façade** (the first coordinator phase since 24 that needed none), `AppCore.shared` greps empty in all five files, and `runWindowCommand` / the façade / `start()` stay put. Moves proven by direct body diff against `HEAD` as in 24 and 25, run **order-preserving rather than sorted** — sorting would hide exactly the reordering AC6 and AC9 exist to catch. Both pinned sequences byte-identical: snippets' guard → early-return → `NSApp.activate` → `confirm` → set flag → `ensureAccessibility()`, and `togglePinned` → `selectClip` → `followToken`. **AC2 not met — 618, not ~560, and the phase doc contradicts itself** the same way 25's did: its "net −195" budget ignores the forwarders it itself mandates (~51 lines) plus 4 properties, so −128 is the ceiling. But **no "no phase owns this" row remains** — of the 618, ~99 are the three palette types **28** extracts by name and ~144 are forwarders **32** deletes, leaving ~375 of composition root (class body alone: 518). `SnippetExpansionCoordinator` gained `unowned core` for the consent dialog, **abandoning 24's deliberate no-`AppCore`-reference shape**; its `showMessage` closure was left in place so every phase-24 body stays byte-identical, leaving two paths to `AppCore` for 29 or 32 to collapse. Debug **and** Release clean, 0 → 0 warnings, all 17 harnesses green, `palette-selection-test` 111,684 unchanged. **`regression.md` waived** — AC5–AC9 structural only, including all three runs the kickoff names (pinned-clip follow, skin-toned emoji, consent-before-Accessibility). **Unblocks 26, 29 and 32**; 29's feature table now needs the four new files named. |
 | 26  | Complete    | `refactor/26-dependency-inversions`        | 2026-08-06 | **M4 closes.** All three inversions gone — `HotKeyManager` takes a `(HotKeyAction) -> String?` resolver (keeping its five static-catalog fallbacks), `collapsedModifierSymbols` is pure, `SystemActionRunner` reports through `onAsyncFailure`; three closures, no protocol, all wired once in `start()`. **AC1 not met — the grep is 6, not 0**: `Core/Backup/BackupActions.swift` is a **fourth** inversion the phase names nowhere (it postdates review §2.3), and two of its six references are the `confirm`/`showNotice` façade every coordinator since 24 deliberately keeps — operator scoped this phase to the three named sites; 29 relocates the file anyway. Two forced deviations: `collapsedModifierSymbols` needs a **fourth** `includesShift:` argument (the doc's three can't reproduce the ✦ set), and `keycaps` reads an injected `@MainActor static` **closure** rather than parameters, because `HotKeyBinding.swift` is must-not-change — a closure, not a snapshot, so a Settings toggle still re-renders every keycap. Debug **and** Release clean, 0 → 0 warnings; the three named harnesses green. Diff +60/−22 against an expected +60/−45; `project.pbxproj` unchanged. **`regression.md` waived** — AC4–AC7 structural only, including the six-owner conflict sweep the kickoff calls the one check that catches a half-wired resolver. **Unblocks 27.** |
-| 27  | Not started |                                            |            |                                                                                                                                                                                                                                        |
+| 27  | Complete    | `refactor/27-extract-designsystem-and-platform` | 2026-08-06 | **M5 opens.** `Core/` 50 → 32 files; `DesignSystem/` (13) and `Platform/` (9) exist. **19 of 20 renames at 100 % similarity**, `EdgeDissolve` and `ThinScrollbar` among them — moved, never opened. The two extractions are byte-identical to their source hunks (`diff` against `HEAD`, not by eye) and both parent files' diffs contain **zero `+` lines**, so no `Theme` token and no cache behaviour changed; `Theme.swift`'s 85 % is the one authorised exception, and the `AppIndex`/`IconCache` pair falls below git's rename threshold so AC2's column cannot show it. `project.yml` needed no edit — XcodeGen derives sources from the directory. **AC4 unmeetable as worded**: `AGENTS.md` holds no harness command lines, so its four stale *paths* and the Project Layout section were fixed instead; the real duplicates live in `docs/snippets.md` and **`ci.yml`**, both updated — a stale path in CI is a red suite, not a stale doc. `checklists/testing.md` updated too (phase 29 requires it; 27 omits it). Debug **and** Release clean, 0 compile warnings either side, `xcodegen` idempotent, all 17 harnesses green under `set -e`, `palette-selection-test` 111,684 unchanged. Comment delta **0**. **`regression.md` waived** — AC7 structural only, including the pixel-identical screenshot pair this phase's whole claim rests on. Pre-existing defect found: `ci.yml`'s `snippets-test` line lacks `HealthTicker.swift` and has been failing since phase 10. **Unblocks 28.** |
 | 28  | Not started |                                            |            |                                                                                                                                                                                                                                        |
 | 29  | Not started |                                            |            |                                                                                                                                                                                                                                        |
 | 30  | Not started |                                            |            |                                                                                                                                                                                                                                        |
